pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "kivy>=2.3.1",
    "pyaudio>=0.2.14",
    "python-can>=4.5.0",
]
